Swap the current document when selecting files
