+++ studio/LonaStudio/Workspace/WorkspaceWindowController.swift
@@ -1,0 +1,14 @@
+//
+//  WorkspaceWindowController.swift
+//  LonaStudio
+//
+//  Created by Devin Abbott on 8/24/18.
+//  Copyright © 2018 Devin Abbott. All rights reserved.
+//
+
+import AppKit
+import Foundation
+
+class WorkspaceWindowController: NSWindowController {
+    
+}
